Removed 4K Sector Size Limit.

+++ inc/winfsp/winfsp.hpp
@@ -492,11 +492,11 @@ public:
     }
 
     /* properties */
-    UINT16 SectorSize()
+    UINT32 SectorSize()
     {
         return _VolumeParams.SectorSize;
     }
-    VOID SetSectorSize(UINT16 SectorSize)
+    VOID SetSectorSize(UINT32 SectorSize)
     {
         _VolumeParams.SectorSize = SectorSize;
     }
